backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
